pre-commit clean
